backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
billiard==3.6.4.0
bit==0.8.0
bitarray==1.2.2
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coincurve==17.0.0
cryptography==37.0.2
cytoolz==0.11.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.50.0
django-constance==2.9.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
eth-abi==2.1.1
eth-account==0.5.7
eth-bloom==1.0.4
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b6
eth-typing==2.3.0
eth-utils==1.10.0
flower==1.2.0
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
hexbytes==0.2.2
humanize==4.4.0
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.6.0
kombu==5.2.4
lru-dict==1.1.7
lxml==4.9.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauthlib==3.2.0
packaging==21.3
parsimonious==0.8.1
phonenumbers==8.12.49
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.20.1
psycopg2-binary==2.9.1
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
rlp==2.0.1
scrypt==0.8.20
semantic-version==2.10.0
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
toolz==0.11.2
tornado==6.2
urllib3==1.26.9
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.29.2
websockets==9.1
wrapt==1.14.1
yarl==1.7.2
